pyproject.toml
[project]
name = "adk"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.22.1",
    "python-dotenv>=1.2.1",
]
